moved rpg-sprite-supplier
