+++ engine/graphics/rpg-sprite-supplier_test.go
@@ -1,4 +1,4 @@
-package engine
+package graphics
 
 import (
 	"retro-carnage/engine/geometry"
